// third_party/dawn/third_party/webgpu-cts/src/external/petamoriken/float16/float16.d.ts
export interface Float16ArrayConstructor {
  readonly prototype: Float16Array;
  new (): Float16Array;
  new (length: number): Float16Array;
  new (elements: Iterable<number>): Float16Array;
  new (array: ArrayLike<number> | ArrayBufferLike): Float16Array;
  new (
    buffer: ArrayBufferLike,
    byteOffset: number,
    length?: number,
  ): Float16Array;

  /**
   * The size in bytes of each element in the array.
   */
  readonly BYTES_PER_ELEMENT: number;

  /**
   * Returns a new array from a set of elements.
   * @param items A set of elements to include in the new array object.
   */
  of(...items: number[]): Float16Array;

  /**
   * Creates an array from an array-like or iterable object.
   * @param elements An iterable object to convert to an array.
   */
  from(elements: Iterable<number>): Float16Array;

  /**
   * Creates an array from an array-like or iterable object.
   * @param elements An iterable object to convert to an array.
   * @param mapfn A mapping function to call on every element of the array.
   * @param thisArg Value of 'this' used to invoke the mapfn.
   */
  from<T>(
    elements: Iterable<T>,
    mapfn: (v: T, k: number) => number,
    thisArg?: any,
  ): Float16Array;

  /**
   * Creates an array from an array-like or iterable object.
   * @param arrayLike An array-like object to convert to an array.
   */
  from(arrayLike: ArrayLike<number>): Float16Array;

  /**
   * Creates an array from an array-like or iterable object.
   * @param arrayLike An array-like object to convert to an array.
   * @param mapfn A mapping function to call on every element of the array.
   * @param thisArg Value of 'this' used to invoke the mapfn.
   */
  from<T>(
    arrayLike: ArrayLike<T>,
    mapfn: (v: T, k: number) => number,
    thisArg?: any,
  ): Float16Array;
}
export declare const Float16Array: Float16ArrayConstructor;